+++ src/markdown/blocks/table.js
@@ -1,37 +1,41 @@
-const { Serializer, Deserializer, Block, BLOCKS, TABLE_ALIGN } = require('../../');
-const reTable = require('../re/table');
+import { Serializer, Deserializer, Block, BLOCKS, TABLE_ALIGN } from '../../';
+import reTable from '../re/table';
 
 /**
  * Deserialize a table with no leading pipe (gfm) to a node.
  * @type {Deserializer}
  */
-const deserializeNoPipe = Deserializer()
-    .matchRegExp(reTable.nptable, (state, match) => {
+const deserializeNoPipe = Deserializer().matchRegExp(
+    reTable.nptable,
+    (state, match) => {
         // Get all non empty lines
         const lines = match[0].split('\n').filter(Boolean);
         const header = lines[0];
-        const align = lines[1];
+        const aligns = lines[1];
         const rows = lines.slice(2);
 
-        const node = parseTable(state, header, align, rows);
+        const node = parseTable(state, header, aligns, rows);
         return state.push(node);
-    });
+    }
+);
 
 /**
  * Deserialize a normal table to a node.
  * @type {Deserializer}
  */
-const deserializeNormal = Deserializer()
-    .matchRegExp(reTable.normal, (state, match) => {
+const deserializeNormal = Deserializer().matchRegExp(
+    reTable.normal,
+    (state, match) => {
         // Get all non empty lines
         const lines = match[0].split('\n').filter(Boolean);
         const header = lines[0];
-        const align = lines[1];
+        const aligns = lines[1];
         const rows = lines.slice(2);
 
-        const node = parseTable(state, header, align, rows);
+        const node = parseTable(state, header, aligns, rows);
         return state.push(node);
-    });
+    }
+);
 
 /**
  * Serialize a table node to markdown
@@ -39,53 +43,44 @@ const deserializeNormal = Deserializer()
  */
 const serialize = Serializer()
     .matchType(BLOCKS.TABLE)
-    .then((state) => {
+    .then(state => {
         const node = state.peek();
         const { data, nodes } = node;
-        const align = data.get('align');
+        const aligns = data.get('aligns');
         const headerRow = nodes.get(0);
         const bodyRows = nodes.slice(1);
 
-        const output = (
-            rowToText(state, headerRow) + '\n'
-            + alignToText(align) + '\n'
-            + bodyRows.map(row => rowToText(state, row)).join('\n')
-            + '\n\n'
-        );
+        const output = `${rowToText(state, headerRow)}\n${alignsToText(
+            aligns
+        )}\n${bodyRows.map(row => rowToText(state, row)).join('\n')}\n\n`;
 
         return state.shift().write(output);
     });
 
-const deserialize = Deserializer()
-    .use([
-        deserializeNoPipe,
-        deserializeNormal
-    ]);
+const deserialize = Deserializer().use([deserializeNoPipe, deserializeNormal]);
 
 /**
  * Parse a table into a node.
  * @param  {State} state
  * @param  {String} headerStr
- * @param  {String} alignStr
+ * @param  {String} alignsStr The line containing the column aligns
  * @param  {String} rowStrs
  * @return {Block} table
  */
-function parseTable(state, headerStr, alignStr, rowStrs) {
+function parseTable(state, headerStr, alignsStr, rowStrs) {
     // Header
     const headerRow = parseRow(state, headerStr);
 
     // Rows
-    const rowTokens = rowStrs.map(rowStr => {
-        return parseRow(state, rowStr);
-    });
+    const rowTokens = rowStrs.map(rowStr => parseRow(state, rowStr));
 
     // Align for columns
-    const alignCells = rowToCells(alignStr);
-    const align = mapAlign(alignCells);
+    const alignsCells = rowToCells(alignsStr);
+    const aligns = mapAligns(alignsCells);
 
     return Block.create({
         type: BLOCKS.TABLE,
-        data: { align },
+        data: { aligns },
         nodes: [headerRow].concat(rowTokens)
     });
 }
@@ -134,10 +129,10 @@ function rowToCells(rowStr) {
     const trimmed = rowStr.trim();
 
     let lastSep = 0;
-    for (let i = 0; i < trimmed.length; i++) {
+    for (let i = 0; i < trimmed.length; i += 1) {
         const prevIdx = i === 0 ? 0 : i - 1;
         const isSep = trimmed[i] === '|';
-        const isNotEscaped = (trimmed[prevIdx] !== '\\');
+        const isNotEscaped = trimmed[prevIdx] !== '\\';
 
         if (isSep && isNotEscaped) {
             // New cell
@@ -161,17 +156,16 @@ function rowToCells(rowStr) {
  * @param {Array<String>}
  * @return {Array<String|null>}
  */
-function mapAlign(align) {
-    return align.map(function(s) {
+function mapAligns(aligns) {
+    return aligns.map(s => {
         if (reTable.alignRight.test(s)) {
             return TABLE_ALIGN.RIGHT;
         } else if (reTable.alignCenter.test(s)) {
             return TABLE_ALIGN.CENTER;
         } else if (reTable.alignLeft.test(s)) {
             return TABLE_ALIGN.LEFT;
-        } else {
-            return null;
         }
+        return null;
     });
 }
 
@@ -184,7 +178,7 @@ function mapAlign(align) {
  */
 function rowToText(state, row) {
     const { nodes } = row;
-    return '| ' + nodes.map(cell => cellToText(state, cell)).join(' | ') + ' |';
+    return `| ${nodes.map(cell => cellToText(state, cell)).join(' | ')} |`;
 }
 
 /**
@@ -200,23 +194,24 @@ function cellToText(state, cell) {
 }
 
 /**
- * Render align of a table to text
+ * Render aligns of a table into a Markdown align row
  *
- * @param {Array<String>} row
+ * @param {Array<String>} aligns
  * @return {String}
  */
-function alignToText(row) {
-    return '|' + row.map(function(align) {
-        if (align == 'right') {
-            return ' ---: |';
-        } else if (align == 'center') {
-            return ' :---: |';
-        } else if (align == 'left') {
-            return ' :--- |';
-        } else {
+function alignsToText(aligns) {
+    return `|${aligns
+        .map(align => {
+            if (align == 'right') {
+                return ' ---: |';
+            } else if (align == 'center') {
+                return ' :---: |';
+            } else if (align == 'left') {
+                return ' :--- |';
+            }
             return ' --- |';
-        }
-    }).join('');
+        })
+        .join('')}`;
 }
 
-module.exports = { serialize, deserialize };
+export default { serialize, deserialize };