+++ HISTORY.md
@@ -5,6 +5,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Unreleased
 
+### 9.0.0
+
+- Rename table aligns data from `align` to `aligns`
+
+### 8.1.3
+
+- Remove warnings when using Slate 0.34.x
+
 ### 8.1.2
 
 - Rename math block to `math_block`